backend/requirements.txt
fastapi>=0.110
uvicorn>=0.29
hello-agents[all]>=0.2.0
pydantic>=2.0
python-dotenv>=1.0
httpx>=0.24
pytest>=8.0
qdrant-client>=1.7
pypdf>=4.0
python-docx>=1.1
